Add files via upload
Clarification: The 3 data sets used are bigger than 25mb.
I download the datasets from this links:
https://www.kaggle.com/datasets/zaheenhamidani/ultimate-spotify-tracks-db
https://www.kaggle.com/datasets/lehaknarnauli/spotify-datasets?select=artists.csv


For my data analysis project, I decided to work with three datasets from Spotify: artists, tracks and Spotify features. My objective was to analyze the data to gain insights into the music industry, and to answer several questions related to music trends and my personal preferences.

Firstly, I began the data preparation process by cleaning and organizing the data sets. This involved removing any duplicate or irrelevant entries, correcting any errors, and ensuring that the data was in a format that could be easily analyzed.

Once the data was prepared, I moved on to the analysis stage. One of the questions I wanted to answer was whether there was a correlation between the popularity of a song or album and its 'energy', a coefficient provided by Spotify. To investigate this, I plotted the data on a heatmap to first reach the conclusion of a positive correlation.

1. Filtered the data set for genre = metal and plotted top 10 artists with most followers:
As part of my analysis of music genres, I decided to focus specifically on the metal genre. I filtered the data set to include only songs and albums with the metal genre tag and created a bar chart showing the top 10 metal artists with the most followers on Spotify.

2. Created a matrix between energy and popularity and reached conclusions based on statistical analysis:
Another question I wanted to explore was whether there was a correlation between a song's energy level and its popularity. To investigate this, I created a matrix with energy level on the x-axis and popularity on the y-axis, with each cell showing the number of songs that fell into that energy/popularity range. I also calculated the Pearson correlation coefficient between energy and popularity to determine the strength of the correlation.

In conclusion, my data analysis project allowed me to gain valuable insights of the music industry and allowed me to gain some experience using pandas and Python.